Fixed compiler warnings /W3

+++ GFX/include/GFX/Resources/Buffer.h
@@ -16,14 +16,14 @@ namespace gfx
     class Buffer
     {
     public:
-        static auto Create(BufferUsage usage, uint64_t size, const void* data = nullptr, bool forceLocalMemory = false) -> OwnedPtr<Buffer>;
-        static auto CreateStaging(uint64_t size, const void* data = nullptr) -> OwnedPtr<Buffer>;
-        static auto CreateVertex(uint64_t size, const void* data = nullptr, bool forceLocalMemory = false) -> OwnedPtr<Buffer>;
-        static auto CreateIndex(uint64_t size, const void* data = nullptr, bool forceLocalMemory = false) -> OwnedPtr<Buffer>;
-        static auto CreateUniform(uint64_t size, const void* data = nullptr) -> OwnedPtr<Buffer>;
+        static auto Create(BufferUsage usage, size_t size, const void* data = nullptr, bool forceLocalMemory = false) -> OwnedPtr<Buffer>;
+        static auto CreateStaging(size_t size, const void* data = nullptr) -> OwnedPtr<Buffer>;
+        static auto CreateVertex(size_t size, const void* data = nullptr, bool forceLocalMemory = false) -> OwnedPtr<Buffer>;
+        static auto CreateIndex(size_t size, const void* data = nullptr, bool forceLocalMemory = false) -> OwnedPtr<Buffer>;
+        static auto CreateUniform(size_t size, const void* data = nullptr) -> OwnedPtr<Buffer>;
 
         virtual ~Buffer() = default;
 
-        virtual void SetData(uint32_t offset, uint32_t size, const void* data) = 0;
+        virtual void SetData(size_t offset, size_t size, const void* data) = 0;
     };
 }
